pyproject.toml
[project]
name = "hiring-confinqa"
version = "0.1.0"
description = "ConfinQA assignment data processing and assessment"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "matplotlib==3.9.2",
  "seaborn==0.13.2",
  "pandas==2.3.0",
  "sympy==1.13.2",
  "pydantic==2.11.7",
  "requests==2.32.3",
  "python-dotenv==0.21.0",
  "langchain-core==0.3.66",
  "langchain-openai==0.3.27",
  "langchain-google-genai==2.1.5",
  "langgraph==0.5.0",
  "numexpr==2.8.7",
  "pymongo==4.13.2",
  "typer==0.12.0",
  "click==8.1.7",
  "ruff==0.12.1",
  "mypy==1.16.1",
  "pylit==0.8.0",
  "streamlit==1.37.0",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
main = "src.main:app"


[tool.ruff]
fix = true
unsafe-fixes = false

[tool.ruff.lint]
select = [
    "F",  # default Ruff rules
    "E",  # pycodestyle Errors
    "W",  # pycodestyle Warnings
    "I",  # import sorting
    "UP", # modern syntax updates
    "B",  # flake8 common pitfalls & best practices
    "Q",  # flake8 consistent quotes
    "N",  # naming conventions,
]

ignore = [
    "B008",
    "D413",
    "E501",
    "W293",
    "W291",
    "PLR2004",
    "D101",
    "D400",

]

fixable = ["ALL"]

extend-select = [
    "PLR2004",
    "D101",
    "D102",
    "D103",
    "D201",
    "D202",
    "D215",
    "T201",
]

[tool.mypy]
disallow_any_generics = true
disallow_untyped_defs = true
disallow_untyped_globals = true
ignore_missing_imports = true
show_error_codes = true
strict = false
warn_return_any = true
plugins = ['pydantic.mypy']
packages = ["src"]
